Show cover art in All Tracks and inline view model
